// aas-web-ui/src/composables/Client/CompanyLookup/types/api.ts
import type { CompanyDescriptor } from './company'

// Todo: types should be autogenerated with every new deployment of the the company lookup service and be provide as a package

export interface PagingMetadata {
  cursor?: string
}

export interface PagedCompanyDescriptors {
  paging_metadata: PagingMetadata
  result: CompanyDescriptor[]
}

export interface Message {
  code?: string
  correlationId?: string
  // Todo: info should be consistent and capitalized in BE
  messageType?: 'Undefined' | 'info' | 'Warning' | 'Error' | 'Exception'
  text?: string
  timestamp?: string
}

export interface CompanyLookupQueryParameters {
  limit?: number
  cursor?: string
  // Todo: search is very stricht and return only with exact match, e.g. if i want to get the company "Fraunhofer IESE" i have to pass the full name
  name?: string
  assetId?: string
}

export interface CompanyLookupResponse<T> {
  success: boolean
  data?: T
  status?: number
  // Note: useRequestHandling is not passing the massages thus we cannot use it here
  // messages?: Message[]
}